refactor(native): extract sidecar preparation
